Split Megatron true-on-policy backend modules
